build.sh

+++ scripts/build.sh
@@ -7,7 +7,9 @@ if [ ! -d ~/.dwm ]; then
 fi
 cp ./scripts/autostart.sh ~/.dwm/autostart.sh
 ln -sf `realpath ./scripts` ~/.dwm/
-cp -f `realpath ./scripts/status_config` ~/.dwm/
+if [ ! -e ~/.dwm/status_config ]; then
+    cp -f `realpath ./scripts/status_config` ~/.dwm/
+fi
 #<}}}
 
 #{{{> .xinitrc
